Updated WorkoutExercise related functionalities; unit tests

+++ src/main/java/sk/krizan/fitness_app_be/service/impl/WorkoutExerciseServiceImpl.java
@@ -1,12 +1,18 @@
 package sk.krizan.fitness_app_be.service.impl;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import sk.krizan.fitness_app_be.controller.exception.ForbiddenException;
 import sk.krizan.fitness_app_be.controller.exception.NotFoundException;
 import sk.krizan.fitness_app_be.controller.request.WorkoutExerciseCreateRequest;
+import sk.krizan.fitness_app_be.controller.request.WorkoutExerciseFilterRequest;
 import sk.krizan.fitness_app_be.controller.request.WorkoutExerciseUpdateRequest;
+import sk.krizan.fitness_app_be.controller.response.PageResponse;
+import sk.krizan.fitness_app_be.controller.response.WorkoutExerciseResponse;
 import sk.krizan.fitness_app_be.model.entity.Exercise;
 import sk.krizan.fitness_app_be.model.entity.User;
 import sk.krizan.fitness_app_be.model.entity.Workout;
@@ -18,6 +24,11 @@ import sk.krizan.fitness_app_be.service.api.ExerciseService;
 import sk.krizan.fitness_app_be.service.api.UserService;
 import sk.krizan.fitness_app_be.service.api.WorkoutExerciseService;
 import sk.krizan.fitness_app_be.service.api.WorkoutService;
+import sk.krizan.fitness_app_be.specification.WorkoutExerciseSpecification;
+import sk.krizan.fitness_app_be.util.PageUtils;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -30,6 +41,34 @@ public class WorkoutExerciseServiceImpl implements WorkoutExerciseService {
     private final WorkoutExerciseRepository workoutExerciseRepository;
 
     private static final String ERROR_NOT_FOUND = "WorkoutExercise with id { %s } does not exist.";
+
+    private static final List<String> supportedSortFields = List.of(
+            WorkoutExercise.Fields.id
+    );
+
+    @Override
+    @Transactional
+    public PageResponse<WorkoutExerciseResponse> filterWorkoutExercises(WorkoutExerciseFilterRequest request) {
+        Specification<WorkoutExercise> specification = WorkoutExerciseSpecification.filter(request);
+        Pageable pageable = PageUtils.createPageable(
+                request.page(),
+                request.size(),
+                request.sortBy(),
+                request.sortDirection(),
+                supportedSortFields
+        );
+        Page<WorkoutExercise> page = workoutExerciseRepository.findAll(specification, pageable);
+        List<WorkoutExerciseResponse> responseList = page.stream()
+                .map(WorkoutExerciseMapper::entityToResponse).collect(Collectors.toList());
+
+        return PageResponse.<WorkoutExerciseResponse>builder()
+                .pageNumber(page.getNumber())
+                .pageSize(page.getSize())
+                .totalElements(page.getTotalElements())
+                .totalPages(page.getTotalPages())
+                .results(responseList)
+                .build();
+    }
 
     @Override
     public WorkoutExercise getWorkoutExerciseById(Long id) {
@@ -47,8 +86,8 @@ public class WorkoutExerciseServiceImpl implements WorkoutExerciseService {
 
     @Override
     @Transactional
-    public WorkoutExercise updateWorkoutExercise(Long id, WorkoutExerciseUpdateRequest request) {
-        WorkoutExercise workoutExercise = getWorkoutExerciseById(id);
+    public WorkoutExercise updateWorkoutExercise(WorkoutExerciseUpdateRequest request) {
+        WorkoutExercise workoutExercise = getWorkoutExerciseById(request.id());
         return workoutExerciseRepository.save(
             WorkoutExerciseMapper.updateRequestToEntity(workoutExercise, request));
     }
@@ -58,13 +97,19 @@ public class WorkoutExerciseServiceImpl implements WorkoutExerciseService {
         User currentUser = userService.getCurrentUser();
         WorkoutExercise workoutExercise = getWorkoutExerciseById(id);
 
+        if (workoutExercise.getWorkout() == null) {
+            throw new RuntimeException("Workout is null");
+        }
+
         if (workoutExercise.getWorkout().getAuthor().getUser() != currentUser
             && !currentUser.getRoleSet().contains(Role.ADMIN)
         ) {
             throw new ForbiddenException();
         }
 
+        workoutExercise.getWorkout().removeFromWorkoutExerciseList(workoutExercise);
         workoutExerciseRepository.delete(workoutExercise);
+
         return workoutExercise.getId();
     }
 }
